refactor(util): fix bad import of getBooleanEnvVar
